requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
deprecation==2.1.0
dj-database-url==2.1.0
Django==4.2.1
django-defender==0.9.7
django-email==0.1.10
django-jet-reboot==1.3.7
django-sendsms==0.5
exceptiongroup==1.2.0
gcloud==0.17.0
googleapis-common-protos==1.62.0
gotrue==2.1.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.6
jws==0.1.3
oauth2client==3.0.0
packaging==23.2
Pillow==10.1.0
postgrest==0.13.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.4.3
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==2.0.1
realtime==1.0.2
redis==5.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.1
typing_extensions==4.9.0
urllib3==2.1.0
websockets==11.0.3
whitenoise==6.6.0
